eu-fiscalizo-backend/requirements.txt
accelerate==1.7.0
altgraph==0.17.4
aniso8601==10.0.1
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
autopep8==2.3.2
beautifulsoup4==4.14.2
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.3
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
dnspython==2.7.0
email_validator==2.2.0
escpos==2.0.0
et_xmlfile==2.0.0
fastapi==0.104.1
filelock==3.18.0
Flask==3.1.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
fsspec==2025.5.0
future==1.0.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
gspread==6.2.0
h11==0.16.0
httplib2==0.22.0
huggingface-hub==0.32.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
linecache2==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
motor==3.3.2
MouseInfo==0.1.3
mpmath==1.3.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.4.0
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyGetWindow==0.0.9
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pymongo==4.15.2
PyMsgBox==2.0.1
PyMySQL==1.1.1
pyparsing==3.2.3
pyperclip==1.11.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-escpos==3.1
python-multipart==0.0.6
pytweening==1.2.0
pytz==2025.2
pyusb==1.3.1
pywhatkit==5.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.2
redis==5.0.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
schedule==1.2.2
secure-smtplib==0.1.1
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.27.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
traceback2==1.4.0
transformers==4.52.3
typing_extensions==4.13.2
tzdata==2025.2
unittest2==1.1.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.24.0
viivakoodi==0.8.0
websockets==12.0
Werkzeug==3.1.3
whatsapp-web==0.0.1
whatsapp-web-py==0.1.0
wikipedia==1.4.0
